TEST: NAV-131 - Move schedule info test to scheduleControllerTest
